+++ components/PostCard/MainPostCard.tsx
