+++ components/PostCard/PostCardSection.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import MainPostCard from '../MainPostCard';
+import MainPostCard from './MainPostCard';
 import { useRouter } from 'next/router';
 import PostCard from '@/components/PostCard/PostCard';
 import { BlogPost } from '@/types/schema';
